third_party/blink/web_tests/http/tests/websocket/cross-origin-expected.txt

WebSocket Cross Origin test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

document.domain=127.0.0.1
ws.url=ws://localhost:8880/origin-test
Connected
PASS origin is "http://127.0.0.1:8000"
PASS successfullyParsed is true

TEST COMPLETE
